refactor: 불필요한 코드 정리, MyPageService -> DashboardService 변경 후 코드 이동
