+++ src/main/java/kosa/server/board/repository/PartyMemberRepository.java
@@ -13,7 +13,6 @@ import java.util.Optional;
 
 @Repository
 public interface PartyMemberRepository extends JpaRepository<PartyMember, Long> {
-    Optional<PartyMember> findByPostAndMember(Post post, Member member);
 
     void deleteAllByPost_Id(Long postId);
     void deleteByPost_IdAndMember_Id(Long postId, Long memberId);
