--- src/main/java/kosa/server/board/service/MyPageService.java
@@ -1,4 +1,0 @@
-package kosa.server.board.service;
-
-public class MyPageService {
-}
